requirements/base.txt
django==4.1.6
django-celery-results
redis==4.5.1
celery==5.2.7
django-celery-beat==2.4.0
gunicorn==20.1.0
uvicorn==0.20.0
requests==2.28.2
djangorestframework==3.14.0
whitenoise==6.3.0
# psycopg2==2.9.5
django-environ==0.9.0
psycopg2-binary